+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,8 +5,6 @@
 import FlutterMacOS
 import Foundation
 
-import flutter_doc_scanner
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
-  FlutterDocScannerPlugin.register(with: registry.registrar(forPlugin: "FlutterDocScannerPlugin"))
 }