+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -6,9 +6,6 @@
 
 #include "generated_plugin_registrant.h"
 
-#include <flutter_doc_scanner/flutter_doc_scanner_plugin_c_api.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
-  FlutterDocScannerPluginCApiRegisterWithRegistrar(
-      registry->GetRegistrarForPlugin("FlutterDocScannerPluginCApi"));
 }